requirements.txt
streamlit>=1.30
pytest>=7.0
anthropic
python-dotenv
